+++ azure_costs_exporter/__init__.py
@@ -1,3 +1,4 @@
 from .main import create_app
+from os import path, getcwd
 
-app = create_app()
+app = create_app(path.join(getcwd(), 'application.cfg'))