+++ azure_costs_exporter/views.py
@@ -1,7 +1,9 @@
 from flask import Blueprint, Response, abort, current_app, request
 from prometheus_client import CONTENT_TYPE_LATEST, CollectorRegistry, generate_latest
 
-from .prometheus_collector import AzureEABillingCollector
+from .enterprise_billing_collector import AzureEABillingCollector
+from .allocated_vm_collector import AzureAllocatedVMCollector
+from .reserved_vm_collector import AzureReservedVMCollector
 
 
 bp = Blueprint('views', __name__)
@@ -15,6 +17,47 @@ def _get_timeout():
         return DEFAULT_SCRAPE_TIMEOUT
 
 
+def _register_billing_collector(registry):
+    timeout = _get_timeout()
+    collector = AzureEABillingCollector(
+        current_app.config['BILLING_METRIC_NAME'],
+        current_app.config['ENROLLMENT_NUMBER'],
+        current_app.config['BILLING_API_ACCESS_KEY'],
+        timeout
+    )
+    registry.register(collector)
+
+
+def _register_allocated_vm_collector(registry):
+    collector = AzureAllocatedVMCollector(
+        current_app.config['APPLICATION_ID'],
+        current_app.config['APPLICATION_SECRET'],
+        current_app.config['AD_TENANT_ID'],
+        current_app.config['SUBSCRIPTION_IDS'],
+        current_app.config['ALLOCATED_VM_METRIC_NAME'],
+    )
+    registry.register(collector)
+
+
+def _register_reserved_vm_collector(registry):
+    collector = AzureReservedVMCollector(
+        current_app.config['APPLICATION_ID'],
+        current_app.config['APPLICATION_SECRET'],
+        current_app.config['AD_TENANT_ID'],
+        current_app.config['RESERVED_VM_METRIC_NAME'],
+    )
+    registry.register(collector)
+
+
+def _register_collectors(registry):
+    if 'BILLING_METRIC_NAME' in current_app.config:
+        _register_billing_collector(registry)
+    if 'ALLOCATED_VM_METRIC_NAME' in current_app.config:
+        _register_allocated_vm_collector(registry)
+    if 'RESERVED_VM_METRIC_NAME' in current_app.config:
+        _register_reserved_vm_collector(registry)
+
+
 @bp.route("/health")
 def health():
     return 'ok'
@@ -22,15 +65,9 @@ def health():
 
 @bp.route("/metrics")
 def metrics():
-    timeout = _get_timeout()
-    collector = AzureEABillingCollector(
-        current_app.config['PROMETHEUS_METRIC_NAME'],
-        current_app.config['ENROLLMENT_NUMBER'],
-        current_app.config['BILLING_API_ACCESS_KEY'],
-        timeout
-    )
     registry = CollectorRegistry()
-    registry.register(collector)
+    _register_collectors(registry)
+
     try:
         content = generate_latest(registry)
         return content, 200, {'Content-Type': CONTENT_TYPE_LATEST}