+++ tests/test_app.py
@@ -1,7 +1,10 @@
 import datetime
+import os
 
 import pytest
 import responses
+import requests
+import mock
 
 from azure_costs_exporter.main import create_app
 from azure_costs_exporter.views import DEFAULT_SCRAPE_TIMEOUT
@@ -18,14 +21,9 @@ def access_key():
     return '3408795poitwiqeotu934t5pqweiut'
 
 
-@pytest.fixture
-def app():
-    return create_app()
-
-
-@pytest.fixture
-def client():
-    application = app()
+def get_client(configuration_name):
+    path = os.path.join(os.getcwd(), 'tests', 'configuration_files', '{}.cfg'.format(configuration_name))
+    application = create_app(path)
     return application.test_client()
 
 
@@ -35,8 +33,8 @@ def now():
 
 
 @responses.activate
-def test_token(client, now, enrollment, access_key):
-
+def test_configured_token_passed_to_billing_collector(now, enrollment, access_key):
+    client = get_client('only_ea_billing')
     responses.add(
         method='GET',
         url="https://ea.azure.com/rest/{0}/usage-report?month={1}&type=detail&fmt=Json".format(enrollment, now),
@@ -49,7 +47,7 @@ def test_token(client, now, enrollment, access_key):
 
 
 @pytest.mark.parametrize('timeout,expected', [('42.3', 42.3), (None, DEFAULT_SCRAPE_TIMEOUT)])
-def test_metrics(app, access_key, now, enrollment, timeout, expected):
+def test_ea_billing_metrics(access_key, now, enrollment, timeout, expected):
     class RequestsMock(responses.RequestsMock):
         def get(self, *args, **kwargs):
             assert kwargs['timeout'] == expected
@@ -67,47 +65,74 @@ def test_metrics(app, access_key, now, enrollment, timeout, expected):
         if timeout is not None:
             headers = {'X-Prometheus-Scrape-Timeout-Seconds': timeout}
 
-        rsp = app.test_client().get('/metrics', headers=headers)
-        url = 'https://ea.azure.com/rest/{enrollment}/usage-report?month={month}&type=detail&fmt=Json'
-        url = url.format(enrollment=enrollment, month=now)
+        rsp = get_client('only_ea_billing').get('/metrics', headers=headers)
         assert rsp.status_code == 200
         assert rsp.data.count(b'azure_costs_eur') == 4
 
 
+def test_failing_target():
+    with mock.patch(
+            "azure_costs_exporter.enterprise_billing_collector.AzureEABillingCollector.collect",
+            side_effect=requests.HTTPError()):
+        rsp = get_client('only_ea_billing').get('/metrics')
+
+        assert rsp.status_code == 502
+        assert rsp.data.startswith(b'Scrape failed')
+
+
 @responses.activate
-def test_metrics_no_usage(app, now, enrollment):
-
-
+def test_allocated_vm_metrics():
+    responses.add(
+        method='POST',
+        url='https://login.microsoftonline.com/tenant_id/oauth2/token',
+        json={"token_type": "Bearer",
+              "expires_in": "3600",
+              "ext_expires_in": "0",
+              "expires_on": "1861920000",
+              "not_before": "1861920000",
+              "resource": "https://management.core.windows.net/",
+              "access_token": "XXXXXX"})
     responses.add(
         method='GET',
-        url="https://ea.azure.com/rest/{0}/usage-report?month={1}&type=detail&fmt=Json".format(enrollment, now),
+        url='https://management.azure.com/subscriptions/SUBSCRIPTION_ID/providers/Microsoft.Compute/virtualMachines?api-version=2017-03-30',
         match_querystring=True,
-        body=api_output_for_empty_months
-    )
+        json={'value': [
+            {
+                'id': '/subscriptions/SUBSCRIPTION_ID/resourceGroups/RESOURCE_GROUP/providers/Microsoft.Compute/virtualMachines/NAME',
+                'location': 'WESTEUROPE',
+                'properties': {'hardwareProfile': {'vmSize': 'SIZE'}}
+            }
+        ]})
 
-    rsp = app.test_client().get('/metrics')
+    rsp = get_client('only_allocated_vm').get('/metrics')
     assert rsp.status_code == 200
-
-    # expect only metric definition and help but no content in the output
-    assert rsp.data.count(b'azure_costs_eur') == 2
+    assert rsp.data.count(b'AZURE_ALLOCATED_VMS') >= 3
 
 
 @responses.activate
-@pytest.mark.parametrize('status', [500, 400])
-def test_failing_target(client, now, status):
+def test_reserved_vm_metrics():
+    responses.add(
+        method='POST',
+        url='https://login.microsoftonline.com/tenant_id/oauth2/token',
+        json={
+            "token_type": "Bearer",
+            "expires_in": "3600",
+            "ext_expires_in": "0",
+            "expires_on": "1861920000",
+            "not_before": "1861920000",
+            "resource": "https://management.core.windows.net/",
+            "access_token": "XXXXXX"})
     responses.add(
         method='GET',
-        url="https://ea.azure.com/rest/{0}/usage-report?month={1}&type=detail&fmt=Json".format(enrollment, now),
+        url='https://management.azure.com/providers/Microsoft.Capacity/reservationOrders?api-version=2017-11-01',
         match_querystring=True,
-        status=status
-    )
+        json={'value': []})  # no reservations sufficient for this test
 
-    rsp = client.get('/metrics')
-
-    assert rsp.status_code == 502
-    assert rsp.data.startswith(b'Scrape failed')
+    rsp = get_client('only_reserved_vm').get('/metrics')
+    assert rsp.status_code == 200
+    assert rsp.data.count(b'AZURE_RESERVED_VMS') >= 4
 
 
-def test_health(client):
-    rsp = client.get('/health')
+def test_health():
+    rsp = get_client('only_ea_billing').get('/health')
     assert rsp.data == b'ok'