+++ CHANGELOG.md
@@ -4,6 +4,14 @@ Change Log
 All notable changes to this project are noted in this file. This project adheres to [Semantic
 Versioning](http://semver.org/).
 
+1.0.0 (unreleased)
+------------------
+
+- Added support for additional metrics: allocated virtual machines and reserved virtual machine
+  instances
+- Renamed configuration option `PROMETHEUS_METRIC_NAME` to `BILLING_METRIC_NAME`
+
+
 0.4.1
 -----